requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
email-validator 
python-multipart
